% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(.file_path)
}
\arguments{
\item{.file_path}{path to the data to import}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Function to import the data
}
